Changed ConfigDbTableModel path to avoid circular import
