+++ pyqt-apps/siriushla/model/__init__.py
@@ -3,3 +3,4 @@
 from .config_names_model import ConfigNamesModel
 from .config_types_model import ConfigTypeModel
 from .pv_configuration_model import PVConfigurationTableModel
+from .config_db_table_model import ConfigDbTableModel
